requirements.txt
mysql-connector-python
pandas
gradio
